tablet: widen the serial type to uint64_t
Internally we still use uint32_t because that's all we get from evdev. But
eventually we'll have 64 bit serials.

+++ src/libinput.c
@@ -1089,7 +1089,7 @@ libinput_tool_get_tool_id(struct libinput_tool *tool)
 	return tool->tool_id;
 }
 
-LIBINPUT_EXPORT uint32_t
+LIBINPUT_EXPORT uint64_t
 libinput_tool_get_serial(struct libinput_tool *tool)
 {
 	return tool->serial;
